input: misc: qti-haptics: check NULL pointer before using it
There might be a rare case that previous play has been terminated and
the next play is generating in VMAX mode so that the play effect pointer
would be set to NULL before the play IRQ being served. So check the
effect pointer before using it in the play IRQ handler.

Change-Id: I2d9ace5d09ffde885b4267e1e78771a169c57bec
Signed-off-by: Fenglin Wu <fenglinw@codeaurora.org>
Signed-off-by: sohamxda7 <sensoham135@gmail.com>
Signed-off-by: ImSpiDy <spidy2713@gmail.com>
Signed-off-by: Souljacker1 <sameerpyo39@gmail.com>